Allow passing in a directory to specfmt
This commit allows `specfmt` to take in a directory, for use as the base in
which to search for the relevant `source` or `*.bs` spec files. In practice, it
expands the usage to support `cargo run ../my-spec-directory`, and specfmt will,
as before, find the source or unique `*.bs` file in that directory instead of
giving up.