Fix handling of path that does not exist
I didn't know how `Path::is_file()` worked. I was assuming that it would return
true if the given path was file-like, even if it did not exist, and therefore
all `Path`s were either `is_file()` or `is_dir()`. This was causing an assert
failure when you pass in a bogus exact file or directory. This changes makes it
so:
 - If the user passes in a file on disk, use it
 - If the user passes in anything else (non-existant file, non-existant
   directory, or an existing directory), then use it as the basis for our spec
   file search. In the case where what was passed in does not exist, we'll hit
   the normal clap failure path.