Fix the `-f/--force` flag
Before this commit, `git_diff()` exclusively retrieved the diff of the base
branch and the current branch, which would always ignore uncommitted changes
even if the `force` flag was supplied. That is, the force flag just got you out
of seeing the uncommitted changes warning, but didn't do anything functionally
beyond that. This PR fixes that.

Fixes #13.