Correctly compute diff
When formatting https://github.com/whatwg/html/pull/9778, I noticed
that the previous branch diff logic never explicitly considered
`current_branch`, and it seemed that in part due to this, the diff was
including (only sometimes though?) merge commits, which is incorrect.

The command line syntax introduced by this commit seems to fix this.